ovn-northd-ddlog: Add ip/ip6 in match condition for ARP request flows.
In C implementation, protocol type is not added, and it relies on
ovn-controller to add the prerequisit in match condition when
translating actions, i.e. ipv4 for arp, and ipv6 for nd_ns.

However, with ddlog, when generating logical flows, if two flows
have same match condition, ddlog will keep updating the SB DB
for both flows interchanging the actions continuously, costing
high CPU. This patch is a workaround to avoid the problem, and it
is not harmful to specify the protocol type clearly.

Root cause still needs to be found in ddlog to avoid other similar
problems, e.g. when conflicting MAC addresses configured, ddlog
will result in same busy-loop behavior.

Signed-off-by: Han Zhou <hzhou8@ebay.com>